+++ CHANGELOG.md
@@ -16,6 +16,8 @@ subsections here on every PR where this is applicable.
 
 ### Added
 
+* Added support to skip staging artifact files during upload by selecting a storage policy by @ibindlish in https://github.com/wandb/wandb/pull/7142
+
 * `run.save()` accepts `pathlib.Path` values; by @timoffex in https://github.com/wandb/wandb/pull/7146
 
 ### Changed